requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
psycopg[binary]>=3.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
requests==2.32.3
PyJWT==2.9.0
passlib
bcrypt
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
watchfiles==0.22.0
pydantic[email]
